Move tests for test util functions to their own file
